Fixing the reload_frequency to reload_interval and changed default
The parameter "reload_frequency" has changed to reload_interval
some time between 0.8 and 1.2.1 This change makes it a consistent
now - there has been and inconsistency - in configuration files
and run_tensorboard.sh it was still RELOAD_FREQUENCY,
 where in the example-app there was
already (unused) RELOAD_INTERVAL enviroment variable.

This commit fixes it and makes it RELOAD_INTERVAL everywhere, together
with fixing the version of tensorboard in the Dockerfile (using latest
in such dockerfile is a bad practice - in case of such incompatible
changes in parameter values, it might simply silently stop working
properly as it did this time).

Also this commit changes the default value of the RELOAD_INTERVAL
parameter. Due to the issue:
https://github.com/tensorflow/tensorboard/issues/158 it seems
that accessing GCS directly causes a lot of costs connected with
high GCP API count usage, therefore if you have thousands of log files
(which is not a lot) it is very easy to overcharge your GCP account with
millions of requests every day just having tensorboard idling and
checking for new data. In our case we got about 4 USD/day for around
3000 files which is quite incredible.